+++ src/components/common/H1/index.module.scss
@@ -1,6 +1,6 @@
 @use '@/styles/mixins' as mixins;
 
-.message {
+.contents {
   margin-bottom: 2rem;
   padding-top: 3.2rem;
   width: 100%;
