+++ src/components/common/index.tsx
@@ -2,3 +2,4 @@ export { default as ErrorNotification } from './ErrorNotification';
 export { default as NotFound } from './NotFound';
 export { default as HomeButton } from './HomeButton';
 export { default as NoBooksYet } from './NoBooksYet';
+export { default as H1 } from './H1';
